refactor: rename component to avoid confusion and improve code readability
Renamed a component to prevent misunderstandings and enhance code clarity.
